Fixed string formatting compiler warnings.

+++ Sources/ReaderThumbRequest.m
@@ -81,7 +81,7 @@
 
 		_fileURL = [url copy]; _password = [phrase copy]; _guid = [guid copy];
 
-		_thumbName = [[NSString alloc] initWithFormat:@"%07d-%04dx%04d", page, w, h];
+		_thumbName = [[NSString alloc] initWithFormat:@"%07d-%04dx%04d", (int)page, (int)w, (int)h];
 
 		_cacheKey = [[NSString alloc] initWithFormat:@"%@+%@", _thumbName, _guid];
